restructure (backend): restructure project
